+++ lib/leftovers/config_loader/keyword_argument_schema.rb
@@ -2,12 +2,12 @@
 
 module Leftovers
   class ConfigLoader
-    class ArgumentPositionSchema < ValueOrObjectSchema
+    class KeywordArgumentSchema < ValueOrObjectSchema
       inherit_attributes_from StringPatternSchema, except: :unless
       attribute :type, ValueOrArraySchema[ValueTypeSchema], require_group: :matcher
       attribute :unless, ValueOrArraySchema[KeywordArgumentSchema], require_group: :matcher
 
-      self.or_value_schema = ScalarArgumentSchema
+      self.or_value_schema = StringSchema
     end
   end
 end