+++ lib/leftovers/matcher_builders/node_pair_key.rb
@@ -1,0 +1,16 @@
+# frozen-string-literal: true
+
+module Leftovers
+  module MatcherBuilders
+    module NodePairKey
+      def self.build(key_matcher)
+        return unless key_matcher
+
+        ::Leftovers::MatcherBuilders::And.build([
+          ::Leftovers::Matchers::NodeType.new(:pair),
+          ::Leftovers::Matchers::NodePairKey.new(key_matcher)
+        ])
+      end
+    end
+  end
+end